+++ frontend/lib/helpers/iconList.ts
@@ -1,5 +1,5 @@
 import { AiOutlineLoading3Quarters } from "react-icons/ai";
-import { BsArrowRightShort } from "react-icons/bs";
+import { BsArrowRightShort, BsChatLeftText } from "react-icons/bs";
 import { CiFlag1 } from "react-icons/ci";
 import {
   FaCheck,
@@ -9,9 +9,12 @@ import {
   FaRegUserCircle,
   FaUnlock,
 } from "react-icons/fa";
-import { FaArrowUpFromBracket } from "react-icons/fa6";
 import { IoIosAdd, IoMdClose, IoMdLogOut } from "react-icons/io";
-import { IoHomeOutline, IoSettingsSharp } from "react-icons/io5";
+import {
+  IoArrowUpCircleOutline,
+  IoHomeOutline,
+  IoSettingsSharp,
+} from "react-icons/io5";
 import { IconType } from "react-icons/lib";
 import {
   LuBrain,
@@ -25,9 +28,9 @@ import {
 } from "react-icons/lu";
 import {
   MdAlternateEmail,
-  MdDelete,
-  MdEdit,
+  MdDeleteOutline,
   MdHistory,
+  MdOutlineModeEditOutline,
   MdUploadFile,
 } from "react-icons/md";
 import { RiHashtag } from "react-icons/ri";
@@ -38,18 +41,19 @@ export const iconList: { [name: string]: IconType } = {
   addWithoutCircle: IoIosAdd,
   brain: LuBrain,
   brainCircuit: LuBrainCircuit,
+  chat: BsChatLeftText,
   check: FaCheck,
   checkCircle: FaCheckCircle,
   chevronDown: LuChevronDown,
   chevronRight: LuChevronRight,
   close: IoMdClose,
   copy: LuCopy,
-  delete: MdDelete,
-  edit: MdEdit,
+  delete: MdDeleteOutline,
+  edit: MdOutlineModeEditOutline,
   email: MdAlternateEmail,
   file: LuFile,
   flag: CiFlag1,
-  followUp: FaArrowUpFromBracket,
+  followUp: IoArrowUpCircleOutline,
   graph: VscGraph,
   hastag: RiHashtag,
   history: MdHistory,